refactor(nim): share one paramiko SSH helper between deploy and teardown
deploy_nim.py and teardown_nim.py each carried a near-identical
ssh_connect()/run_cmd() pair, and the timeout-aware variant only lived in
teardown_nim, so the two had drifted. Extract a single timeout-aware
implementation into shared/ssh_paramiko.py and import it from both scripts, so
SSH fixes apply to both and deploy_nim gains the same timeout diagnostics.
